Convert last remaining charts
