+++ front/app/containers/Admin/dashboard/users/Charts/PieChartByCategory.tsx
@@ -1,9 +1,8 @@
 // libraries
-import React from 'react';
+import React, { useRef } from 'react';
 
 // intl
-import { injectIntl, FormattedMessage } from 'utils/cl-intl';
-import { WrappedComponentProps } from 'react-intl';
+import { FormattedMessage, useIntl } from 'utils/cl-intl';
 import messages from '../../messages';
 
 // utils
@@ -11,7 +10,6 @@ import renderTooltip from './renderPieChartByCategoryTooltip';
 import { roundPercentages } from 'utils/math';
 
 // styling
-import { withTheme } from 'styled-components';
 import { categoricalColorScheme } from 'components/admin/Graphs/styling';
 
 // components
@@ -28,31 +26,18 @@ import {
 import PieChart from 'components/admin/Graphs/PieChart';
 
 // resources
-import GetSerieFromStream from 'resources/GetSerieFromStream';
 
 // typings
-import { IStreamParams, IStream } from 'utils/streams';
 import { LegendItem } from 'components/admin/Graphs/_components/Legend/typings';
 import { IGraphFormat } from 'typings';
-import { IUsersByBirthyear } from 'api/users_by_birthyear/types';
 import { IUsersByCustomField } from 'api/users_by_custom_field/types';
-
-interface DataProps {
-  serie: IGraphFormat;
-}
+import useUsersByCustomField from 'api/users_by_custom_field/useUsersByCustomField';
 
 interface Serie extends IGraphFormat {
   percentage: number;
 }
 
-type ISupportedDataType = IUsersByBirthyear | IUsersByCustomField;
-
-interface InputProps {
-  stream: (
-    streamParams?: IStreamParams | null,
-    customId?: string
-  ) => IStream<ISupportedDataType>;
-  convertToGraphFormat: (data: ISupportedDataType) => IGraphFormat | null;
+interface Props {
   startAt: string | null | undefined;
   endAt: string | null;
   currentGroupFilter: string | undefined;
@@ -62,27 +47,51 @@ interface InputProps {
   className?: string;
   customId: string;
   xlsxEndpoint: string;
+  id: string;
 }
 
-interface Props extends InputProps, DataProps {}
+const PieChartByCategory = ({
+  startAt,
+  endAt,
+  className,
+  graphTitleString,
+  xlsxEndpoint,
+  currentGroupFilter,
+  currentGroupFilterLabel,
+  id,
+}: Props) => {
+  const { formatMessage } = useIntl();
+  const currentChart = useRef<any>();
+  const { data: usersByCustomField } = useUsersByCustomField({
+    start_at: startAt,
+    end_at: endAt,
+    group: currentGroupFilter,
+    id,
+  });
 
-class PieChartByCategory extends React.PureComponent<
-  Props & WrappedComponentProps
-> {
-  currentChart: React.RefObject<any>;
+  const convertCheckboxToGraphFormat = (data: IUsersByCustomField) => {
+    const {
+      series: { users },
+    } = data.data.attributes;
+    const res = ['_blank', 'true', 'false'].map((key) => ({
+      value: users[key] || 0,
+      name: formatMessage(messages[key]),
+      code: 'key',
+    }));
 
-  constructor(props: Props & WrappedComponentProps) {
-    super(props);
-    this.currentChart = React.createRef();
-  }
+    return res.length > 0 ? res : null;
+  };
 
-  makeLegends = (row, i): LegendItem => ({
+  const serie =
+    usersByCustomField && convertCheckboxToGraphFormat(usersByCustomField);
+
+  const makeLegends = (row, i): LegendItem => ({
     icon: 'circle',
     color: categoricalColorScheme({ rowIndex: i }),
     label: `${row.name} (${row.percentage}%)`,
   });
 
-  addPercentages = (serie): Serie | undefined => {
+  const addPercentages = (serie): Serie | undefined => {
     if (!serie) return;
     const percentages = roundPercentages(serie.map((row) => row.value));
     return serie.map((row, i) => ({
@@ -91,80 +100,57 @@ class PieChartByCategory extends React.PureComponent<
     }));
   };
 
-  render() {
-    const {
-      startAt,
-      endAt,
-      className,
-      graphTitleString,
-      serie,
-      xlsxEndpoint,
-      currentGroupFilter,
-      currentGroupFilterLabel,
-    } = this.props;
+  const percentagesSerie = addPercentages(serie);
 
-    const percentagesSerie = this.addPercentages(serie);
-
-    return (
-      <GraphCard className={className}>
-        <GraphCardInner>
-          <GraphCardHeader>
-            <GraphCardTitle>{graphTitleString}</GraphCardTitle>
-            {serie && (
-              <ReportExportMenu
-                name={graphTitleString}
-                svgNode={this.currentChart}
-                xlsx={{ endpoint: xlsxEndpoint }}
-                currentGroupFilter={currentGroupFilter}
-                currentGroupFilterLabel={currentGroupFilterLabel}
-                startAt={startAt}
-                endAt={endAt}
-              />
-            )}
-          </GraphCardHeader>
-          {!percentagesSerie ? (
-            <NoDataContainer>
-              <FormattedMessage {...messages.noData} />
-            </NoDataContainer>
-          ) : (
-            <PieChartStyleFixesDiv>
-              <PieChart
-                data={percentagesSerie}
-                width={164}
-                mapping={{
-                  angle: 'value',
-                  name: 'name',
-                }}
-                pie={{
-                  startAngle: 0,
-                  endAngle: 360,
-                  outerRadius: 60,
-                }}
-                tooltip={renderTooltip()}
-                legend={{
-                  items: percentagesSerie.map(this.makeLegends),
-                  maintainGraphSize: true,
-                  marginLeft: 50,
-                  position: 'right-center',
-                }}
-                innerRef={this.currentChart}
-              />
-            </PieChartStyleFixesDiv>
+  return (
+    <GraphCard className={className}>
+      <GraphCardInner>
+        <GraphCardHeader>
+          <GraphCardTitle>{graphTitleString}</GraphCardTitle>
+          {serie && (
+            <ReportExportMenu
+              name={graphTitleString}
+              svgNode={currentChart}
+              xlsx={{ endpoint: xlsxEndpoint }}
+              currentGroupFilter={currentGroupFilter}
+              currentGroupFilterLabel={currentGroupFilterLabel}
+              startAt={startAt}
+              endAt={endAt}
+            />
           )}
-        </GraphCardInner>
-      </GraphCard>
-    );
-  }
-}
+        </GraphCardHeader>
+        {!percentagesSerie ? (
+          <NoDataContainer>
+            <FormattedMessage {...messages.noData} />
+          </NoDataContainer>
+        ) : (
+          <PieChartStyleFixesDiv>
+            <PieChart
+              data={percentagesSerie}
+              width={164}
+              mapping={{
+                angle: 'value',
+                name: 'name',
+              }}
+              pie={{
+                startAngle: 0,
+                endAngle: 360,
+                outerRadius: 60,
+              }}
+              tooltip={renderTooltip()}
+              legend={{
+                items: percentagesSerie.map(makeLegends),
+                maintainGraphSize: true,
+                marginLeft: 50,
+                position: 'right-center',
+              }}
+              innerRef={currentChart}
+            />
+          </PieChartStyleFixesDiv>
+        )}
+      </GraphCardInner>
+    </GraphCard>
+  );
+};
 
-const PieChartByCategoryWithHoCs = injectIntl(
-  withTheme(PieChartByCategory as any) as any
-);
-
-const WrappedPieChartByCategory = (inputProps: InputProps) => (
-  <GetSerieFromStream {...inputProps}>
-    {(serie) => <PieChartByCategoryWithHoCs {...serie} {...inputProps} />}
-  </GetSerieFromStream>
-);
-
-export default WrappedPieChartByCategory;
+export default PieChartByCategory;
